Update 40_hw_assembly_db18.md
Just adding a note for people make the same mistake as I do. If every reader will know the difference between standoffs and spacers, maybe don't add this. 
(What a spacer is is also well visible in the parts-picture, but in the picture below the changed line, it does look as if it could be spacers.)