Added OS X workaround
For the unsupported OS X version of dts to run, you need to remove that installed flag, otherwise it will not download the commands properly. 

I'm not sure if it is best to put it here as a note, or in the git documentation linked to in that same section, but it should be visible somewhere.